rename distance_utils.py
