+++ tests/unit/algorithms/common/utils/test_dist_utils.py
@@ -1,12 +1,12 @@
 """Tests for test_distance.py"""
-from otx.algorithms.common.utils import distance_utils
-from otx.algorithms.common.utils.distance_utils import append_dist_rank_suffix
+from otx.algorithms.common.utils import dist_utils
+from otx.algorithms.common.utils.dist_utils import append_dist_rank_suffix
 from tests.test_suite.e2e_test_system import e2e_pytest_unit
 
 
 @e2e_pytest_unit
 def test_append_dist_rank_suffix_not_distributed_training(mocker):
-    mock_os = mocker.patch.object(distance_utils, "os")
+    mock_os = mocker.patch.object(dist_utils, "os")
     mock_os.environ = {}
     file_name = "temporary.pth"
     new_file_name = append_dist_rank_suffix(file_name)
@@ -16,8 +16,8 @@ def test_append_dist_rank_suffix_not_distributed_training(mocker):
 
 @e2e_pytest_unit
 def test_append_dist_rank_suffix_distributed_training(mocker):
-    mock_os = mocker.patch.object(distance_utils, "os")
-    mock_os.environ = {"LOCAL_RANK": "2"}
+    mock_os = mocker.patch.object(dist_utils, "os")
+    mock_os.environ = {"LOCAL_RANK" : "2"}
     new_file_name = append_dist_rank_suffix("temporary.pth")
 
     assert new_file_name == "temporary_proc2.pth"
